requirements.txt
crewai==0.5.0
openai>=1.7.1
python-dotenv>=1.0.0
langchain>=0.1.0
moonshotai>=0.0.18